Refactor Conf's middleware in to separate packages and add a i18n middleware template.
